+++ .github/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 Update <small>_ January 2024</small>
 
+- feat: auto-generation of the command table (24/01/2024)
 - feat: migrate reportedIssues command (22/01/2024)
 - feat: docSearch command (22/01/2024)
 - fix: vatsim events max fields (20/01/2024)